requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
powerline-status==2.7
PyJWT==2.7.0
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
